Refactored inventory service to move inventory snapshot methods into their own service; also moved dashboard-related methods into dashboard service
